refactor: move to download/mutations
